lint
SCCS-vsn: sys/netinet/in_var.h 6.5
SCCS-vsn: sys/netinet/ip_input.c 6.23
SCCS-vsn: sys/vax/vax/Locore.c 6.10
SCCS-vsn: sys/vax/vax/machdep.c 6.30
SCCS-vsn: sys/kern/kern_clock.c 6.17
SCCS-vsn: sys/kern/uipc_usrreq.c 6.23
SCCS-vsn: sys/vax/if/if_ddn.c 6.6
SCCS-vsn: sys/vax/if/if_dmc.c 6.13
SCCS-vsn: sys/vax/if/if_hdh.c 6.7
SCCS-vsn: sys/vax/if/if_pcl.c 6.8
SCCS-vsn: sys/vax/if/if_qe.c 6.3
SCCS-vsn: sys/vax/if/if_vv.c 6.21
SCCS-vsn: sys/vax/uba/ct.c 6.6
SCCS-vsn: sys/netns/ns.h 6.12
SCCS-vsn: sys/netns/ns_error.c 6.8
SCCS-vsn: sys/netns/ns_if.h 6.4
SCCS-vsn: sys/netns/ns_input.c 6.13
SCCS-vsn: sys/netns/ns_pcb.c 6.8
SCCS-vsn: sys/netns/ns_pcb.h 6.6

+++ usr/src/sys/netinet/ip_input.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)ip_input.c	6.22 (Berkeley) %G%
+ *	@(#)ip_input.c	6.23 (Berkeley) %G%
  */
 
 #include "param.h"
@@ -849,7 +849,7 @@ ip_forward(ip, ifp)
 	 * Save at most 64 bytes of the packet in case
 	 * we need to generate an ICMP message to the src.
 	 */
-	mcopy = m_copy(dtom(ip), 0, imin(ip->ip_len, 64));
+	mcopy = m_copy(dtom(ip), 0, imin((int)ip->ip_len, 64));
 
 	sin = (struct sockaddr_in *)&ipforward_rt.ro_dst;
 	if (ipforward_rt.ro_rt == 0 ||
